Standardize name and position for "sienv" variable
Summary:
In a previous diff in this stack, we discovered that the use of the variable name "env" to refer to the symbol index environment was confusing.  We typically use the variable name "env" to mean "ServerEnv.env".

Let's rename it to `sienv` everywhere.

We'll also doublecheck to make sure `sienv` is always the first parameter for all symbol index methods, and that it is always a named parameter.

Reviewed By: young-k

Differential Revision: D15976770

fbshipit-source-id: 0a2c68796e72f05c0b4a349d75c703c0a3fd5874