Add env to symbol index calls
Summary:
This diffstack enables all references used by the symbol index to be held in a single environment object.

This diff extends the signature of symbol index methods to take, as input, the environment object.  It's marked as ref in cases where it is modified, and non-ref when it will not be changed.

It should always be the first parameter in a symbol index call.

Note that the use of the variable name "env" is problematic: we often use that to mean ServerEnv.env.  Hmm.  Let's fix that in the next diff in the stack.

Reviewed By: arxanas

Differential Revision: D15974643

fbshipit-source-id: b3275a7428e0ba067d9b11dfbb7bdcfb608d01ca