Remove Typing_taccess.env.ids
Summary:
Typing_taccess expands a type access which can be a series of type consts. Handling multiple accesses makes the code in Typing_taccess needlessly complex and it can be greatly simplified by focusing on expanding only a single type const.

That is what this diff does. It removes the ids field from `Typing_taccess.env` and has the code only operate on a single id at a time. In `Typing_phase` when we localize a `Taccess` we fold over the id list.

Reviewed By: vassilmladenov

Differential Revision: D15617225

fbshipit-source-id: f41e04f58a91cad18ee902917bd5d0cc14e38421