Use root as this_ty in Typing_taccess
Summary: When expanding a type constant it may contain a `this` type. We were not properly setting it. This fixes it so it is set to the root type that we are expanding. We set this when `dep_tys` is empty since that means we have encountered an expression dependent or generic type for the first time. Subsequent generic types we see should not update the this type given the current implementation. This will be changed in a later diff in this stack.

Reviewed By: vassilmladenov

Differential Revision: D15617241

fbshipit-source-id: 73e545c5f66fd3933406a32435c4dcd8d0ea8bff