Eliminated trail from Typing_taccess.env
Summary:
The trail was used as part of tracking the reason of the expansion. Overall the reasons we produced for type access were too verbose and generally not helpful. I simplified the reasons and the subsequent code in Typing_taccess, by handling the complexity of merging reasons in Typing_reason.to_string.

I also fixed a case I ran into after {D15781671} where we would report (Expect T...But got T). Whenever this arises its due to exact types so I added to the error message in that case.

Reviewed By: CatherineGasnier

Differential Revision: D15954310

fbshipit-source-id: f52efd27f767556186cfbc419598f98ff7349ddf