Add ArrayData::nvGet{Int,Str}Pos
Summary: Add functions to the ArrayData API that let you go from a int or string key to a pos. They complement the nvGetKey/atPos APIs that let you go from a pos to a key/value at that pos. If the key is not present in the array they return the 'canonical invalid position' (what iter_end() returns).

Reviewed By: jjgriego

Differential Revision: D15813921

fbshipit-source-id: 46a3eb8d567a10d751dbfb2b0e87f5d96113086f