Make array_intersect_key faster when left input is larger.
Summary:
We can make array_intersect_key faster when the left input is larger than the right input by iterating the right input and probling the left one, accumulating positions from the left input that have matching keys in the right input, and creating the output in a separate pass over the left input and the accumulated positions.

I tried a similar approach for array_diff_key (P66328777 P66328904) but micro-benchmarking showed that, while we could get speedups using this technique when there was minimal overlap in keys between the left and right inputs, when there was maximal overlap (i.e. when all the keys in the smaller input were also in the larger one) this change typically performed worse. Since we can't know ahead of time what the overlap is, we can't effectively make use of this approach for array_diff_key.

Reviewed By: paulbiss

Differential Revision: D15813929

fbshipit-source-id: 4841c99f4d31bb45a4c9f4f92b2995f49c600365