Rename confusingly named variable in ext_array.cpp
Summary:
A bunch of functions in ext_array.cpp have a local named isKey. It's not actually set based on whether anything was a key - they're checking if the input isArrayLikeType and then using it to decide whether to coerce intish strings to ints. The functions it's passed to take a parameter named convertIntLikeStrs, a much better name.
Rename the isKey boolean to convertIntLikeStrs (also inverting the value that's stored in it to match the new name).

Reviewed By: jjgriego

Differential Revision: D15813876

fbshipit-source-id: 923e900709630bf519592bc8189d5d263ecbeed6