Unified "initialize" method for symbol index
Summary:
As part of the project to replace all global refs for the symbol index, we'd like to have a single unified "initialize" function that constructs the index.  This would mean that we could construct separate symbol indexes that would work side by side, or have the option to swap one symbol index for another at runtime.

The new syntax looks like this:
```
 env.local_symbol_table := (SymbolIndex.initialize
      ~quiet:genv.local_config.ServerLocalConfig.symbolindex_quiet
      ~provider_name:genv.local_config.ServerLocalConfig.symbolindex_search_provider
      ~namespace_map
      ~savedstate_file_opt:genv.local_config.ServerLocalConfig.symbolindex_file
      ~workers:genv.workers);
```

Reviewed By: arxanas

Differential Revision: D15977386

fbshipit-source-id: 5f662619269afb8298e21f69a728020c97eb13e3