Move make_next_files from ServerInitCommon to ServerFiles
Summary:
This function is useful for making batch tools that need not be built into `hh_server`.
Moving it out into `server_utils` makes it unnecessary to link to `server_extras`, which brings in a ton of dependencies, including the search service. The linking time with `server_extras` is measured in minutes, while linking with `server_utils` instead should take a few seconds.

Differential Revision: D16021889

fbshipit-source-id: 68d69aaca1d2a1832f213dcdbd21213548e07de9