Always parse files on-demand for `Ast_provider` with local-memory backend
Summary:
We shouldn't need to keep parse trees in memory since they should be accessed only once. After the first access, we should cache any decls in memory. If the decls fall out of the cache, then the right thing to do is to go back to disk to re-parse the file.

Keeping parse trees in memory is more useful as a performance optimization for bulk typechecking. For the serverless IDE, it's much better to reduce memory usage instead.

Reviewed By: pittsw

Differential Revision: D15893915

fbshipit-source-id: c51d86741684a7a8fdba6ae8afa7e6cdae8ae815