Pass a TAST into `ServerHover` to provide hover results
Summary:
The current `ServerHover.go` code relies on implicit global state. In particular, it expects the currently-operated-on file to be available in the file heap, keyed by the magic empty path (`ServerIdeUtils.path`).

In the serverless IDE, when backed by the `Provider_config.Local_memory` provider, the file heap isn't a special heap. All accesses are always passed through to the filesystem, so you can't insert this special magical key-value pair for IDE requests.

Consequently, `textDocument/hover` fails to work upon the switch to `Local_memory`. The principled fix is to pass through all the information that's needed for `ServerHover` to work on the document currently open in the editor (which may not even have been saved to disk).

This commit accomplishes that, based on the same technique as used in {D15059410}.

Reviewed By: pittsw

Differential Revision: D15893920

fbshipit-source-id: 30549efe75c80947e874049ebb5a2d95a10069fd