Better chunk size for parallel
Summary:
The default chunk size is 120; but that means that for small inputs,
parallel::for_each ends up not parallelizing much (if at all).

For the most part that doesn't matter too much, but I was adding some
ad-hoc parallelization of groups of unrelated things, and everything
ran in a single thread.

Reviewed By: binliu19, alexeyt

Differential Revision: D16022889

fbshipit-source-id: 1475b006d7a3d1eced1e9988f5f755b7fadada8e