Parallelize index destruction
Summary:
It takes a long time. We used to hide some of that cost by
parallelizing it with the unit emitter; but thats no longer safe.

We can move destruction of the dependencyMap (which takes most of the
time) to run in parallel with the final pass (its not used once the
index is frozen).

Parallelizing the rest saves a few more seconds.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D16022888

fbshipit-source-id: cb0dd4381e93d12c086b217b5e7d94e5d3e55ad2