Fix handling of stack overflows
Summary:
When we are in prologue, ActRec is no longer considered pre-live. As long as we
tell unwinder the locals are already freed, unwinder is perfectly capable of
unwinding it, so just do that and directly throw FatalErrorException.

With the existence of new style FCall* opcodes, it is incorrect to consider
ActRec in prologue as pre-live, as there's no FPI region to inform unwinder to
handle it specially.

Reviewed By: paulbiss

Differential Revision: D16034914

fbshipit-source-id: 832b3b0155a7597db998cfc3a61352109da0b431