Remove `ServerHover.go` in favor of `ServerHover.go_ctx`
Summary:
Use `ServerHover.go_ctx` everywhere so that we can clean up the legacy code path.

`go_ctx` is superior because it relies less on global caches to do its work; almost everything can be accessed using the `ctx` passed in. Note that it currently will read files from disk using the `File_provider`, but we should change it so that it also has to go through `ctx` to access the filesystem as well.

Reviewed By: pittsw, young-k

Differential Revision: D15933194

fbshipit-source-id: 01463e270457b8c205ea985a20f36046e4a1b9a7