Lock objects when unwinding from constructors.
Summary:
When we unwind from a constructor that's called as part of a new expression (as opposed to one called via e.g. parent::__construct()), the LockObj bytecode we emitted after the FCallCtor will not run, but we still need the object to be locked so that we don't end up with mutable instances floating around in the event the constructor stashed it somewhere.

In HHBBC, mark the FCA of FCallCtor bytecodes with the ConstructNoConst bit if we may optimize away the LockObj that comes after (and therefore also the copyof the object being constructed on the top of the caller's stack).

In the unwinder, when unwinding both PHP and C++ exceptions, add a check after we've torn down a frame to see if that frame came from an FCallCtor bytecode and that that bytecode's FCA is not marked with ConstructNoConst by hhbbc. If so lock the instance on the top of the caller's stack, then continue unwinding.

Reviewed By: markw65

Differential Revision: D15939210

fbshipit-source-id: ab977744b3c22e344c902dca08dde7b7fa9e311b