FPushObjMethod: use interpOne() instead of PUNT()
Summary:
PUNT() is dangerous, as we first trace, guarding the stack input, PUNT()ing
based on the stack input type, which triggers retrace, which remembers to go to
interpreter directly, not using the stack input type, so guard is relaxed.

Then we form a region, thinking that we don't need to guard, but we reach the
JIT implementation of the opcode that tries to popC(), which fails on an
assertion because the only thing we know is that it is TGen.

This is not causing any known issues right now, but will cause issues with the
next diff that merges FPushObjMethod opcodes with FCall.

Reviewed By: paulbiss

Differential Revision: D16043080

fbshipit-source-id: f3ecd05ef8961e2b6209b0cf2d19c1c6ee29d8f0