Fix "type" field in debug_backtrace() when inlined
Summary:
If we inline a function containing a call to debug_backtrace, and
elide the frame-pointer, we'll encounter a "fake" ActRec when creating
the backtrace. Such a fake always has the locals dec-reffed flag set,
so we'll always give it a type of "::", even when it should be
"->". Use isStatic() on the func instead of checking for the presence
of $this in the frame.

Reviewed By: paulbiss

Differential Revision: D16033405

fbshipit-source-id: c69b09676f71accb50996daf0381e7ffdd28abdb