HHVM Debugger: Fix segfault in fbdbg REPL context
Summary: This was a regression caused by D15890730. In the REPL context for fbdbg, there is no frameId so the debugger's FrameObject pointer is nullptr. Later, the current code tries to deference that to determine the frame depth at which we're evaluating to find its variable environment context. It should be using 0 in this case, since we're in the top frame of the dummy thread in the REPL context.

Reviewed By: alexeyt

Differential Revision: D16057138

fbshipit-source-id: f4b9e12190464a0115351e4bd999538dda75c9fd