ZOMG Fast Docblocks
Summary:
In the process of capturing a docblock for a symbol, we currently use a method called X.  We use that method because we don't know the "kind" of the symbol.

That method requires a fully typed AST to determine the kind.  Fully typed ASTs are slow.  Bummer!

But wait ... there's more.  Since, in the `completionItem/resolve` pathway, we already know the kind, we can convert it and bypass the need for typing.

Let's add a second pathway that already knows the kind and can avoid having to generate types!

Notes that have to be completed before this can be merged:
* There is a LOT of logging cruft in this code that needs to be cleaned out.
* I'm not sure if the ServerRpc DOCBLOCK_AT works.  I need to check that, or, failing that, maybe I should remove that server rpc pathway completely.
* I would bet money that the other pathways that call for docblocks already know the kind of a symbol as well.  They could probably be optimized just as nicely.

Reviewed By: 2BitSalute, pittsw

Differential Revision: D16044718

fbshipit-source-id: 7eb3ff4c17117c7be28442f3ad4631276082e246