Typing: defer class declaration in `Typing_classes_heap` if the type checker option is set
Summary:
Today, when we type check file `A.php`, we declare dependencies as we encounter them. This leads to bad type checking times on files that depend on a lot of classes. This is relevant in the case of generated files, for example.

The idea of this diff stack is to defer declaring dependencies and discard type errors for `A.php` if its dependencies weren't already available.

=====

Here, `Typing_classes_heap` uses the `defer_class_declaration` type checker option to refuse to declare class dependencies on-demand. This results in type errors as if the class really didn't exist.

Differential Revision: D15869497

fbshipit-source-id: e8d8472177dfe54db4fb520fb38205b8b879da7b