Apply equalities before doing is_sub_type
Summary: Currently, `simplify_subtype` does not "look" at bounds on type variables, even when used by `is_sub_type` to determine if a subtype holds for *any* instantiation of type variables. In future, we should support this, but for now, let's apply equalities implied by bounds before calling `is_sub_type`. (This is a prerequisite for the fix to arithmetic overloading.)

Reviewed By: kmeht

Differential Revision: D16048812

fbshipit-source-id: c728fcd6137dc414c745c23a173c3866c4a9dc63