Fix hhbbc build time regression
Summary:
With FPush/FCall, we record failed attempts to flatten, so that we
don't keep trying over and over.

Merging FPushObjMethod* into FCallObjMethod* dropped that, and as a
result we ended up spending 60+% of the time *just* adding
dependencies when we failed to resolve due to inlining depth (so
obviously a lot more time interpreting the three levels of inlining to
get there).

Reviewed By: alexeyt

Differential Revision: D16068962

fbshipit-source-id: 81c9cd580916fc47c050a2e86ee9b170da3be09b