Fix positions of Awaitall nodes produced by parsing function bodies
Summary:
Consider the following example (added as a test case below):
```
//// file1.php
<?hh

async function gen(): Awaitable<void> {}

//// file2.php
<?hh

function call((function(): void) $f): void {
  $f();
}

function test(): void {
  call(() ==> await gen());
}
```
It produces the following cryptic error message:
```
File "", line 0, characters 0-0:
PRIMARY ERROR POSITION IS NOT IN CURRENT FILE: please fix (NastCheck[3003])
File "", line 0, characters 0-0:
await can only be used inside async functions
```
The problem here is that the body of the lambda passed to `call` is parsed as (or rewritten to) an `Awaitall` statement with the position `Pos.none`, which is then used by `Naming_coroutine_check` when reporting an error.  The fix is to use the position of the original expression (before rewriting).

Reviewed By: andrewjkennedy

Differential Revision: D16071723

fbshipit-source-id: 38ace46cbe6fee67c9612215d006770437f09696