Handle DefLabels while adjusting BC markers during InlineReturn optimization
Summary:
adjustBCMarkers() during InlineReturn frame optimization was
assuming the parent FP came from either a DefFP or a DefInlineFP. The
third possibility is a DefLabel.

It needs to retrieve the stack pointer offset of the parent FP. To
handle DefLabels, walk one of the DefLabels sources and recursively
try to obtain its offset. This will continue until we hit either a
DefFP or a DefInlineFP. All inputs to the DefLabel should ultimately
resolve to the same offset (in debug builds we'll now verify
this). Also check for self-referential DefLabels, which we shouldn't
get with frame pointers right now.

Reviewed By: markw65

Differential Revision: D16070538

fbshipit-source-id: b100be6af26e76b23858c1721df56c8c48cd4ac3