Add FuncCred bytecode
Summary:
Adds a bytecode to push a resource onto the stack containing a pointer to the currently executing function.

Down the road this resource will be passed into an HHVM extension that will be able to make a Crypto Auth Token for this function. In this context it can be thought of as "proof that this function was executed and desires to make a Crypto Auth Token"

Reviewed By: paulbiss

Differential Revision: D15488979

fbshipit-source-id: 6a6bc8b9c691db7c28251ddb778932b0c4bf4b63