Re-abstract `Typing_classes_heap.Api.t`
Summary:
This type should be abstract, because it exposes whether the class type is accessed using an "eager" or "lazy" class type. The "lazy" class type is used to implement shallow decls. People can get confused about what API to use, especially since the type is currently not-abstract.

The type was made un-abstract in {D15404424} to support duplicating some of the logic in `Decl_provider`, except that the logic shouldn't depend on the caches. My first attempt copied the code, but also had to make the type no longer abstract so that it could access the internals (as OCaml has no "friend" visibility). That actually is a step backward for abstraction, since the type is now exposed, even if the duplicated logic isn't.

The approach in this commit is to expose the duplicated logic instead, and re-abstract the type. This is still suboptimal, but should be less confusing to consumers, since references to the function won't leak through the codebase like references to the type could.

Differential Revision: D15989104

fbshipit-source-id: 7b9f0e34c46ccde137143bf065dbf906119f6a84