Various speed-ups for vasm-graph-color
Summary:
Perform various architectural speed-ups for vasm-graph-color:

- Drop the requirement that physical registers have to be
  SSAized. This required changing a lot of the algorithm to deal with
  physical registers directly. However, this leads to a variety of
  speedups. Namely, there's no longer any need for pseudo
  instructions, so they can be removed entirely, along with the passes
  to create them and remove them. We also no longer need precolored
  Vregs, nor any of the constrained instruction logic.

- Drop the recursive recoloring optimization phase. This optimization
  is very effective, but also very slow. It's also not very compatible
  with physical registers not being SSAized. It also requires the
  calculation of a variety of metadata (like dominator tree
  calculation), which can all be dropped. This diff doesn't have a
  replacement for this optimization, but the next diff will (to
  minimize changes as much as possible).

- Add infrastructure to cache instruction operands. Rather than using
  a vasm visitor to get operands over and over, cache the operands
  into VregSets once and just return that. This means that you need to
  invalidate the cache when you actually change the operands.

- Short-cut the spiller when possible. A simple register pressure
  calculation over the unit can tell us the blocks where we might have
  to generate a spill. For the blocks where we know we'll never spill,
  the spiller can do a much simpler calculation over the entire block
  at once. In the (common) case where no block will generate a spill,
  we can bypass the spiller entirely.

- Do some quick short-cuts in phi lowering to avoid expensive logic.

Overall these changes speed up the algorithm by ~4x on a lot of common
workloads, more so in some special cases.

As mentioned above, these changes impact the colorer pass a
lot. However, the colorer pass is also reworked to use a new copy
coalescing heuristic. I want to keep the size of this diff down as
much as possible, so I decided to postpone the colorer changes to the
next diff (which implements the new heuristic). This means the colorer
does not really work on this diff and is just a placeholder. Without
the new coalescing heuristic, it would produce very bad code anyways.

Sorry for the size and complexity of this diff, but all of the changes
are very inter-linked and was hard to factor out.

Reviewed By: mxw, mofarrell

Differential Revision: D15864849

fbshipit-source-id: 59678d5f9a7decdf95563720a0d0db357b61cb88