Fix a segfault due to magic getters/setters
Summary:
The jit tries to optimize away uses of TVRef during minstr
sequences. It can't do that if a magic getter/setter is involved,
unless it knows there's a real property that can't have been unset.

The check for a real property failed to check that it was accessible,
so the TVRef was optimized away, and then we crashed trying to
dereference nullptr.

Reviewed By: paulbiss, alexeyt

Differential Revision: D16097766

fbshipit-source-id: b6274624964eb1c7dffe7ed1b85cc144c6ab9067