Avoid realpath calls for autoloader files (with runtime option)
Summary: We can state that the autoloader always uses realpath using the `Eval.TrustAutoloaderPath` runtime option. This allows us to skip all of the expensive realpath checks. This does not apply to the PHP callback which autoloads files and may account for a large set of calls. A future native autoloader could reduce the cost even further.

Reviewed By: markw65

Differential Revision: D15578375

fbshipit-source-id: ab4c2594c9c80db47d1e320ce048a470252c1e20