Use optional shape fields of type nothing to represent unset fields
Summary: When calling `Shapes::removeKey($s, 'f')`, instead of adding `'f'` to the set of unset shape fields in the type of `$s`, we can change the type of `'f'` to be an optional shape field of type `nothing`.  (Such a field is either not set in the shape or it is set to a value of type `nothing`, but there are no such values, which leaves us with the first alternative.)  This allows us to get rid of the unset fields set in the shape type.  In this diff, I make sure that we never use the unset fields set (because it should always be empty anyway) and that when we need to construct a new unset fields set, we just use an empty set.  In the next diff I'm going to get rid of the unset fields entirely and rename the constructors of `shape_fields_known` to `Closed_shape` and `Open_shape` (and the type itself to `shape_kind`).

Reviewed By: andrewjkennedy

Differential Revision: D16090796

fbshipit-source-id: 3e79b5776aa386f1e6220ab72210ca35bb44e4d1