Native AutoloadMap using Watchman
Summary:
This is a Facebook-only implementation of a native AutoloadMap based on Watchman.

At the moment, the biggest limitation is that you need to call `HH\autoload_watchman_query` to use it. For each request, we need to know the repository root involved. In order to get rid of this userspace function, I'll probably need to implement something which infers the repo root by crawling the repo tree for a `.watchmanconfig` or `.hhconfig` file.

There may also be some WWW cleanup necessary to remove the userspace function. WWW's initialization code branches on whether or not a given symbol has already been loaded, and that concept doesn't make sense if we have an autoloader ready on line 1 of every PHP request.

Reviewed By: paulbiss

Differential Revision: D14395412

fbshipit-source-id: 79c058ab9d83fba7f5fb0d1a518933d889ff8aa8