prevent early xenon SIGPROF from terminating process
Summary: When Xenon is enabled with a timer, the first timer event is set to
occur at a random interval up to the configured interval. If the timer signals
before block_sync_signals_and_start_handler_thread is invoked, the unhandled
SIGPROF terminates the process. Fixed by purposely ignoring SIGPROF until the
correct handler is installed. (Credit to @enjmusic for the fix.)