Add switch to disable legacy soft typehints
Summary: As a part of the migration to the new `@` attribute syntax, soft typehints are being migrated to the `__Soft` attribute. This diff makes the legacy syntax a parse error (behind the `--disable-legacy-soft-typehints` switch), thus ensuring that soft typehints are all using the new attribute to make way for the attribute syntax migration.

Reviewed By: hgoldstein

Differential Revision: D16100609

fbshipit-source-id: 89cb272de497c865e652517321eac3111a057c3b