Enable access to $this from fb_setprofile handler
Summary: The current implementation of `fb_setprofile` passes `this_ptr`, the numeric pointer value of `$this`, to the registered handler. It does not appear this can be used to access properties of the object. I have introduced an additional array key in the handler's data parameter, `this_obj`, which provides access to the object on which the method is invoked, in a similar fashion to `fb_intercept`.

Reviewed By: paulbiss

Differential Revision: D16059618

fbshipit-source-id: b7bb3e6c3b9c57de8667d421879f2088d9989bfa