Don't report unbound constant error when resolving function names
Summary: If the receiver of a function call expression is a name, it is always treated by HHVM as the identifier of a global function. HHVM will not try to resolve it to a constant, so there is no ambiguity and the type checker should not try to look it up as a constant either.

Reviewed By: paulbiss

Differential Revision: D16140345

fbshipit-source-id: 2cf7e332ed7b54302a23f506198f346969f41d4a