Classify fun('nonexistent') as a new error code
Summary: `fun()` may in the future default to emitting function pointers (set `Eval.EmitFuncPointers=1` to test this). As a result, it will need to resolve the function name, and will fatal if no corresponding function is found. Make a unique message and code for this error so that it stands out.

Reviewed By: paulbiss

Differential Revision: D16140560

fbshipit-source-id: d0be0918577bb45ee9a0a41cff85782400314a9f