Fix crash with interfaces and abstract classes
Summary:
Abstract classes don't have a vtableVec, but you can call static
methods on them, meaning that the lsb class will be abstract (and have
no vtableVec). If we try to do vtable dispatch, and the
InstanceOfIfaceVtable gets optimized to an InstanceOfBitmask, we'll
crash trying to load the Func from the vtable.

Mark InstanceOfIfaceVtable which might have this problem as
not-optimizable (we could actually optimize them down to a check for
nullptr - but that can be a followup).

Reviewed By: ricklavoie

Differential Revision: D16144198

fbshipit-source-id: f15da9a7a63c6ad8dbb6fafcfe4b068bc9b98ddb