Refine expression to type of null at conditions
Summary:
Putting this diff up again, because since we now have proper refinements of disjunctions in D9923073, it would be nice if the following worked:
```
function f(mixed $x): void {
  if ($x is string || $x === null) {
    expect_nullable_string($x);
  }
}
```

When we encounter a condition that is a null check, we refine the type of the variable to nonnull on one side. This diff adds refinement of the variable to type of null on the other side.

Example code:
```
function get(): ?SomeClass;

$x = get(); // ?SomeClass
if ($x !== null) {
  // $x now has type SomeClass (refined)
} else {
  // pre: $x has type ?SomeClass (not refined)
  // now: $x has type null
}
```

This can be useful when the type of the variable changes in one branch and we don't get into incorrect type after the conditional check. e.g.
```
function get(): ?SomeClass;

$x = get(); // ?SomeClass
if ($x !== null) {
  // $x now has type SomeClass (refined)
  $x = new OtherClass();
  // $x now has type OtherClass
} else {
  // do nothing
}
// pre: $x has type (?SomeClass | OtherClass)
// post: $x has type ?OtherClass
```

Reviewed By: Wilfred

Differential Revision: D8399732

fbshipit-source-id: 05d7be2c4f831d4b38efe89b05a2a3148de26f5a