refinement of conjunctions and disjunctions
Summary:
The following example would previously be accepted:
```
function f(bool $b, string $x): void {
  if ($b || ($x = 1) !== 0) {
    expect_string($x);
  }
}
```
but if `$b` is false, `$x` is an int, even in the `else` block, so `expect_string` here should produce an error.

So I fixed it, and by fixing it, I also made refinements of disjunctions work in the general case, e.g.:
```
if ($x is int || $x is string) {
  // $x is arraykey here
}
```
The reason why I needed this is that in the `else` branch in the first example, the condition is equivalent to `!($b && ($x = 1) === 0)`, i.e. `!$b || ($x = 1) !== 0`, i.e. in effect a disjunction.

Reviewed By: andrewjkennedy

Differential Revision: D9923073

fbshipit-source-id: 48f03aeea4ef7f245fa27d97d587d8aa5bf61698