prevent early xenon SIGPROF from terminating process (#8527)
Summary:
When Xenon is enabled with a timer, the first timer event is set to
occur at a random interval up to the configured interval. If the timer signals
before block_sync_signals_and_start_handler_thread is invoked, the unhandled
SIGPROF terminates the process. Fixed by purposely ignoring SIGPROF until the
correct handler is installed. (Credit to enjmusic for the fix.)
Pull Request resolved: https://github.com/facebook/hhvm/pull/8527

Reviewed By: binliu19

Differential Revision: D16150112

fbshipit-source-id: 7b385ce159b17fe6fe6fd5fe6acbb85fc0b75b08