Mangle into UnitCache whether the file is .hhas when AllowHhas is set
Summary: When `AllowHhas=1` and we pass an hhas file without `.hhas` extension, we'd treat it as PHP file. This is fine. But if we change the name of the file to `filename.hhas` but not the contents, it would have the same sha1 hence we'd pull it from the cache treating it as identical. But this time, we do need it treat it as an hhas file.

Reviewed By: markw65

Differential Revision: D16153438

fbshipit-source-id: cc7c779a935d098d249cac6bd4ed888d893c4bf8