Improve error messages generated by subtyping checks with an option on the left and union on the right
Summary: The subtyping constraint `?t <: t1 | t2 | ... | tN` is simplified to a conjunction `(null <: t1 | t2 | ... | tN) && (t <: t1 | t2 | ... | tN)`.  If the first constraint is unsatisfiable the user will see an error of the form "Expected `t1 | t2 | ... | tN` but got `null`", which is not great, especially in the case when `?t = ?nonnull  = mixed`.  To improve the error message, let us report a mismatch between the original types `?t` and `t1 | t2 | ... | tN` when we fail to satisfy the first constraint, `null <: t1 | t2 | ... | tN`.

Reviewed By: CatherineGasnier

Differential Revision: D16148853

fbshipit-source-id: 6e8af51d27d3c4c22f8a63413596523d3fa7fccb